requirements.txt
boto3==1.17.84
botocore==1.20.84
certifi==2020.11.8
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dataclasses==0.8
Flask==2.0.1
Flask-Cors==3.0.10
idna==2.10
importlib-metadata==4.3.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
zipp==3.4.1
